data-challenge-api/requirements.txt
fastapi
uvicorn
urllib
google-cloud-storage
pydantic
